✔️ Score incrementation
Used formula to calculate how many points for an invader depending on row.
This could let us scale the game up for more rows later if we really wanted.
But really I just didn't want to hardcode values.

Added helper function to pad ints with leading 0s and used that for all
3 score outputs.